feat: move sign up to use auth hook
